[SCALA] New endpoint added, category and opinion repository moved to product repository to easier management

+++ shop-backend/app/models/ProductTag.scala
@@ -2,7 +2,7 @@ package models
 
 import play.api.libs.json._
 
-case class ProductTag(id: Int, productId: Int, text: String)
+case class ProductTag(id: Int, productId: Long, text: String)
 
 object ProductTag {
   implicit val opinionFormat: OFormat[ProductTag] = Json.format[ProductTag]
